+++ libnavigation-core/src/main/java/com/mapbox/navigation/core/internal/extensions/MapboxNavigationObserverEx.kt
@@ -55,4 +55,6 @@ class MapboxNavigationObserverChain : MapboxNavigationObserver {
     fun clear() {
         queue.clear()
     }
+
+    fun toList(): List<MapboxNavigationObserver> = queue.toList()
 }