+++ libnavui-dropin/src/test/java/com/mapbox/navigation/dropin/ViewBinderTest.kt
@@ -4,6 +4,8 @@ import com.mapbox.navigation.base.ExperimentalPreviewMapboxNavigationAPI
 import com.mapbox.navigation.dropin.ActionButtonDescription.Position.END
 import com.mapbox.navigation.dropin.ActionButtonDescription.Position.START
 import com.mapbox.navigation.dropin.binder.EmptyBinder
+import com.mapbox.navigation.dropin.binder.infopanel.InfoPanelBinder
+import com.mapbox.navigation.dropin.binder.infopanel.MapboxInfoPanelBinder
 import com.mapbox.navigation.ui.base.lifecycle.UIBinder
 import io.mockk.mockk
 import org.junit.Assert.assertEquals
@@ -51,6 +53,7 @@ internal class ViewBinderTest {
                 infoPanelContentBinder = UIBinder.USE_DEFAULT
                 actionButtonsBinder = UIBinder.USE_DEFAULT
                 customActionButtons = emptyList()
+                infoPanelBinder = InfoPanelBinder.defaultBinder()
             }
         )
 
@@ -62,6 +65,7 @@ internal class ViewBinderTest {
         assertTrue(sut.infoPanelContentBinder.value == null)
         assertTrue(sut.actionButtonsBinder.value == null)
         assertTrue(sut.customActionButtons.value.isEmpty())
+        assertTrue(sut.infoPanelBinder.value is MapboxInfoPanelBinder)
     }
 
     private fun customization() = ViewBinderCustomization().apply {
@@ -77,5 +81,6 @@ internal class ViewBinderTest {
             ActionButtonDescription(mockk(), START),
             ActionButtonDescription(mockk(), END)
         )
+        infoPanelBinder = mockk()
     }
 }